Move progress bar dialog to ElephantDatasetMixin
- move a progress bar dialog from RabbitMQService to ElephantDatasetMixin
