+++ src/main/java/org/elephant/actions/RabbitMQService.java
@@ -26,14 +26,8 @@
  ******************************************************************************/
 package org.elephant.actions;
 
-import java.awt.BorderLayout;
 import java.io.IOException;
 import java.util.concurrent.TimeoutException;
-
-import javax.swing.JDialog;
-import javax.swing.JLabel;
-import javax.swing.JProgressBar;
-import javax.swing.SwingConstants;
 
 import org.apache.commons.lang.exception.ExceptionUtils;
 import org.elephant.actions.ElephantStatusService.ElephantStatus;
@@ -42,8 +36,6 @@ import org.elephant.actions.mixins.ElephantStateManagerMixin;
 import org.elephant.actions.mixins.WindowManagerMixin;
 import org.scijava.listeners.Listeners;
 
-import com.eclipsesource.json.Json;
-import com.eclipsesource.json.JsonObject;
 import com.rabbitmq.client.AlreadyClosedException;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
@@ -76,14 +68,15 @@ public class RabbitMQService extends AbstractElephantService
 
 	private final Listeners.List< RabbitMQStatusListener > rabbitMQStatusListeners;
 
-	private final ExceptionHandler emptyExceptionHandler = new EmptyExceptionHander();
+	private final Listeners.List< RabbitMQDatasetListener > rabbitMQDatasetListeners;
 
-	private final ProgressDialog progressDialog = new ProgressDialog();
+	private final ExceptionHandler emptyExceptionHandler = new EmptyExceptionHander();
 
 	public RabbitMQService()
 	{
 		super();
 		rabbitMQStatusListeners = new Listeners.SynchronizedList<>();
+		rabbitMQDatasetListeners = new Listeners.SynchronizedList<>();
 	}
 
 	public void start()
@@ -145,22 +138,7 @@ public class RabbitMQService extends AbstractElephantService
 		channel.basicConsume( RABBITMQ_QUEUE_UPDATE, true, callbackUpdate, consumerTag -> {} );
 		channel.queueDeclare( RABBITMQ_QUEUE_DATASET, false, false, false, null );
 		final DeliverCallback callbackDataset = ( consumerTag, delivery ) -> {
-			final String message = new String( delivery.getBody(), "UTF-8" );
-			final JsonObject jsonObject = Json.parse( message ).asObject();
-			final int tMax = jsonObject.get( "t_max" ).asInt();
-			final int tCurrent = jsonObject.get( "t_current" ).asInt();
-			if ( tCurrent < tMax )
-			{
-				progressDialog.setProgressBarValue( 100 * tCurrent / tMax );
-				progressDialog.setLabelText( String.format( "%d / %d", tCurrent, tMax ) );
-				progressDialog.setVisible( true );
-			}
-			else
-			{
-				progressDialog.setProgressBarValue( 0 );
-				progressDialog.setLabelText( "" );
-				progressDialog.setVisible( false );
-			}
+			rabbitMQDatasetListeners.list.forEach( l -> l.messageDelivered( consumerTag, delivery ) );
 		};
 		channel.basicConsume( RABBITMQ_QUEUE_DATASET, true, callbackDataset, consumerTag -> {} );
 	}
@@ -187,6 +165,11 @@ public class RabbitMQService extends AbstractElephantService
 	public Listeners< RabbitMQStatusListener > rabbitMQStatusListeners()
 	{
 		return rabbitMQStatusListeners;
+	}
+
+	public Listeners< RabbitMQDatasetListener > rabbitMQDatasetListeners()
+	{
+		return rabbitMQDatasetListeners;
 	}
 
 	private class EmptyExceptionHander implements ExceptionHandler
@@ -225,40 +208,4 @@ public class RabbitMQService extends AbstractElephantService
 		{}
 
 	}
-
-	private class ProgressDialog extends JDialog
-	{
-		private static final long serialVersionUID = 1L;
-
-		private final JProgressBar progressBar = new JProgressBar();
-
-		private final JLabel lblText = new JLabel( "" );
-
-		public ProgressDialog()
-		{
-			setDefaultCloseOperation( JDialog.DO_NOTHING_ON_CLOSE );
-			setTitle( "Gnerate Dataset" );
-
-			progressBar.setValue( 0 );
-			progressBar.setStringPainted( true );
-			getContentPane().add( progressBar, BorderLayout.CENTER );
-
-			lblText.setHorizontalAlignment( SwingConstants.CENTER );
-			getContentPane().add( lblText, BorderLayout.NORTH );
-
-			pack();
-			setLocationRelativeTo( null );
-		}
-
-		public void setLabelText( final String text )
-		{
-			lblText.setText( text );
-		}
-
-		public void setProgressBarValue( final int value )
-		{
-			progressBar.setValue( value );
-		}
-	}
-
 }
